finish handle hall of fame
